organize package-related files under lib/
